requirements.txt
couleur==0.5.0
misaka==1.0.2
python-termstyle==0.1.20
PyYAML==3.10
